test(health): move health endpoint tests into routes/health
Adds a _tests.rs sibling per file: logic_tests.rs covers build()
directly, http_tests.rs and mcp_tests.rs hold the GET /health and MCP
health tool tests that used to live in the top-level routes/*_tests.rs
files, next to the handlers they exercise.
