+++ src/routes/health/http.rs
@@ -10,3 +10,7 @@ use logic::HealthResponse;
 pub async fn health(State(state): State<crate::routes::http::AppState>) -> Json<HealthResponse> {
     Json(logic::build(state.uptime_start))
 }
+
+#[cfg(test)]
+#[path = "http_tests.rs"]
+mod http_tests;
