+++ src/routes/http_tests.rs
@@ -317,43 +317,6 @@ async fn build_app_serves_machines() {
 }
 
 #[tokio::test]
-async fn build_app_serves_health() {
-    let app = build_app(crate::routines::new_store());
-    let resp = app
-        .oneshot(
-            Request::builder()
-                .uri("/api/v1/health")
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-    assert_eq!(resp.status(), StatusCode::OK);
-    let bytes = axum::body::to_bytes(resp.into_body(), usize::MAX)
-        .await
-        .unwrap();
-    let json: serde_json::Value = serde_json::from_slice(&bytes).unwrap();
-    assert_eq!(json["status"], "ok");
-    assert_eq!(json["running"], true);
-    // The dependencies section reports tmux presence so the UI/CLI can flag a missing dependency.
-    assert!(
-        json["dependencies"]["tmux"].is_boolean(),
-        "health payload should carry a boolean dependencies.tmux flag, got: {json}"
-    );
-    // Likewise for python3, which the built-in `claude` agent's setup step depends on (#404).
-    assert!(
-        json["dependencies"]["python3"].is_boolean(),
-        "health payload should carry a boolean dependencies.python3 flag, got: {json}"
-    );
-    assert_eq!(json["version"], env!("CARGO_PKG_VERSION"));
-    // The resolved machine name is surfaced so clients can identify which daemon answered.
-    assert!(
-        json["machine"].is_string() && !json["machine"].as_str().unwrap().is_empty(),
-        "health payload should carry a non-empty machine name, got: {json}"
-    );
-}
-
-#[tokio::test]
 async fn build_app_serves_ui_at_root() {
     let app = build_app(crate::routines::new_store());
     let resp = app
